refactor: rename pkg/chaos/docker to pkg/chaos/lifecycle (issue 6)
Lifecycle commands (stop/kill/pause/restart/remove/exec) are runtime-agnostic
and depend only on container.Lifecycle/Lister/Executor interfaces — no Docker
imports. The 'docker' name predated containerd and podman runtimes and was
misleading. Mechanical rename via git mv plus package declaration and import
path updates.
