+++ pkg/chaos/lifecycle/cmd/kill.go
@@ -6,7 +6,7 @@ import (
 	"fmt"
 
 	"github.com/alexei-led/pumba/pkg/chaos"
-	"github.com/alexei-led/pumba/pkg/chaos/docker"
+	"github.com/alexei-led/pumba/pkg/chaos/lifecycle"
 	"github.com/urfave/cli"
 )
 
@@ -23,7 +23,7 @@ func NewKillCLICommand(ctx context.Context) *cli.Command {
 			cli.StringFlag{
 				Name:  "signal, s",
 				Usage: "termination signal, that will be sent by Pumba to the main process inside target container(s)",
-				Value: docker.DefaultKillSignal,
+				Value: lifecycle.DefaultKillSignal,
 			},
 			cli.IntFlag{
 				Name:  "limit, l",
@@ -53,7 +53,7 @@ func (cmd *killContext) kill(c *cli.Context) error {
 	// get limit for number of containers to kill
 	limit := c.Int("limit")
 	// init kill command
-	killCommand, err := docker.NewKillCommand(chaos.DockerClient, params, signal, limit)
+	killCommand, err := lifecycle.NewKillCommand(chaos.DockerClient, params, signal, limit)
 	if err != nil {
 		return fmt.Errorf("could not create kill command: %w", err)
 	}
