+++ pkg/chaos/lifecycle/pause.go
@@ -1,4 +1,4 @@
-package docker
+package lifecycle
 
 import (
 	"context"
