+++ pkg/chaos/lifecycle/pause_test.go
@@ -1,4 +1,4 @@
-package docker
+package lifecycle
 
 import (
 	"context"
