+++ pkg/chaos/lifecycle/remove.go
@@ -1,4 +1,4 @@
-package docker
+package lifecycle
 
 import (
 	"context"
