+++ pkg/chaos/lifecycle/restart_test.go
@@ -1,9 +1,10 @@
-package docker
+package lifecycle
 
 import (
 	"context"
 	"errors"
 	"testing"
+	"time"
 
 	"github.com/alexei-led/pumba/pkg/container"
 	"github.com/stretchr/testify/assert"
@@ -11,17 +12,16 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
-func TestRemoveCommand_Run(t *testing.T) {
+func TestRestartCommand_Run(t *testing.T) {
 	type wantErrors struct {
-		listError   bool
-		removeError bool
+		listError    bool
+		restartError bool
 	}
 	type fields struct {
 		names   []string
 		pattern string
-		force   bool
-		links   bool
-		volumes bool
+		labels  []string
+		timeout time.Duration
 		limit   int
 		dryRun  bool
 	}
@@ -38,32 +38,39 @@ func TestRemoveCommand_Run(t *testing.T) {
 		errs     wantErrors
 	}{
 		{
-			name: "remove matching containers by names",
+			name: "restart matching containers by names",
 			fields: fields{
-				names: []string{"c1", "c2", "c3"},
-				force: true,
+				names:   []string{"c1", "c2", "c3"},
+				timeout: 1 * time.Second,
 			},
 			args:     args{ctx: context.TODO()},
 			expected: container.CreateTestContainers(3),
 		},
 		{
-			name: "remove matching containers by filter with limit",
+			name: "restart matching labeled containers by names",
+			fields: fields{
+				names:   []string{"c1", "c2", "c3"},
+				labels:  []string{"key=value"},
+				timeout: 1 * time.Second,
+			},
+			args:     args{ctx: context.TODO()},
+			expected: container.CreateLabeledTestContainers(3, map[string]string{"key": "value"}),
+		},
+		{
+			name: "restart matching containers by filter with limit",
 			fields: fields{
 				pattern: "^c?",
-				force:   true,
-				links:   true,
+				timeout: 1 * time.Second,
 				limit:   2,
 			},
 			args:     args{ctx: context.TODO()},
 			expected: container.CreateTestContainers(3),
 		},
 		{
-			name: "remove random matching container by names",
+			name: "restart random matching container by names",
 			fields: fields{
 				names:   []string{"c1", "c2", "c3"},
-				force:   true,
-				links:   true,
-				volumes: true,
+				timeout: 1 * time.Second,
 			},
 			args:     args{ctx: context.TODO(), random: true},
 			expected: container.CreateTestContainers(3),
@@ -71,59 +78,62 @@ func TestRemoveCommand_Run(t *testing.T) {
 		{
 			name: "no matching containers by names",
 			fields: fields{
-				names: []string{"c1", "c2", "c3"},
+				names:   []string{"c1", "c2", "c3"},
+				timeout: 1 * time.Second,
 			},
 			args: args{ctx: context.TODO()},
 		},
 		{
 			name: "error listing containers",
 			fields: fields{
-				names: []string{"c1", "c2", "c3"},
+				names:   []string{"c1", "c2", "c3"},
+				timeout: 1 * time.Second,
 			},
 			args:    args{ctx: context.TODO()},
 			wantErr: true,
 			errs:    wantErrors{listError: true},
 		},
 		{
-			name: "error removing container",
+			name: "error restarting container",
 			fields: fields{
-				names: []string{"c1", "c2", "c3"},
+				names:   []string{"c1", "c2", "c3"},
+				timeout: 1 * time.Second,
 			},
 			args:     args{ctx: context.TODO()},
 			expected: container.CreateTestContainers(3),
 			wantErr:  true,
-			errs:     wantErrors{removeError: true},
+			errs:     wantErrors{restartError: true},
 		},
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
 			mockClient := container.NewMockClient(t)
-			k := &removeCommand{
+			k := &restartCommand{
 				client:  mockClient,
 				names:   tt.fields.names,
 				pattern: tt.fields.pattern,
-				force:   tt.fields.force,
-				links:   tt.fields.links,
-				volumes: tt.fields.volumes,
+				labels:  tt.fields.labels,
+				timeout: 1 * time.Second,
 				limit:   tt.fields.limit,
 				dryRun:  tt.fields.dryRun,
 			}
+			opts := container.ListOpts{Labels: tt.fields.labels}
 			if tt.errs.listError {
-				mockClient.EXPECT().ListContainers(mock.Anything, mock.AnythingOfType("container.FilterFunc"), mock.MatchedBy(func(opts container.ListOpts) bool { return opts.All == true })).Return(nil, errors.New("ERROR"))
+				mockClient.EXPECT().ListContainers(mock.Anything, mock.AnythingOfType("container.FilterFunc"), opts).Return(nil, errors.New("ERROR"))
 			} else {
-				mockClient.EXPECT().ListContainers(mock.Anything, mock.AnythingOfType("container.FilterFunc"), mock.MatchedBy(func(opts container.ListOpts) bool { return opts.All == true })).Return(tt.expected, nil)
+				mockClient.EXPECT().ListContainers(mock.Anything, mock.AnythingOfType("container.FilterFunc"), opts).Return(tt.expected, nil)
 				if tt.expected != nil {
-					removeCall := mockClient.EXPECT().RemoveContainer(mock.Anything, mock.AnythingOfType("*container.Container"), tt.fields.force, tt.fields.links, tt.fields.volumes, tt.fields.dryRun)
-					if tt.errs.removeError {
-						removeCall.Return(errors.New("ERROR")).Once()
+					restartCall := mockClient.EXPECT().RestartContainer(mock.Anything, mock.AnythingOfType("*container.Container"), tt.fields.timeout, tt.fields.dryRun)
+					if tt.errs.restartError {
+						restartCall.Return(errors.New("ERROR")).Once()
 					} else if tt.args.random {
-						removeCall.Return(nil).Once()
+						restartCall.Return(nil).Once()
 					} else {
 						count := len(tt.expected)
 						if tt.fields.limit > 0 && tt.fields.limit < count {
 							count = tt.fields.limit
 						}
-						removeCall.Return(nil).Times(count)
+						restartCall.Return(nil).Times(count)
 					}
 				}
 			}
